+++ apps/vue-2/.eslintrc.cjs
@@ -3,13 +3,4 @@ const { eslintrc } = require('@mgdis/linting-vue2');
 module.exports = {
   root: true,
   ...eslintrc,
-  rules: {
-    ...eslintrc.rules,
-    'vuejs-accessibility/no-redundant-roles': [
-      'error',
-      {
-        main: ['main'],
-      },
-    ],
-  },
 };