+++ packages/mg-components/src/components/molecules/inputs/mg-input-password/tests/mg-input-password.e2e.ts
@@ -193,4 +193,30 @@ test.describe('mg-input-password', () => {
       await expect(page.locator('.e2e-screenshot')).toHaveScreenshot();
     });
   });
+
+  test('Should reset value and error when calling reset method', async ({ page }) => {
+    const html = createHTML({ ...baseArgs });
+    await page.setContent(html);
+
+    await page.locator('mg-input-password.hydrated').waitFor();
+
+    // Enter a value that doesn't match the pattern
+    await page.locator('mg-input-password input').fill('Chase');
+
+    // Set an error message intentionally
+    await page.evaluate(() => {
+      document.querySelector('mg-input-password').setError(false, "Message d'erreur de test");
+    });
+
+    // Check state with value and error
+    await expect(page.locator('.e2e-screenshot')).toHaveScreenshot();
+
+    // Call reset method
+    await page.evaluate(() => {
+      document.querySelector('mg-input-password').reset();
+    });
+
+    // Check that the input has been reset and the error has been removed
+    await expect(page.locator('.e2e-screenshot')).toHaveScreenshot();
+  });
 });