+++ packages/mg-components/src/components/molecules/inputs/mg-input-select/mg-input-select.tsx
@@ -316,6 +316,19 @@ export class MgInputSelect {
   }
 
   /**
+   * Reset value, validity and error state
+   */
+  @Method()
+  async reset(): Promise<void> {
+    if (!this.readonly) {
+      this.value = null;
+      this.checkValidity();
+      this.errorMessage = undefined;
+      this.hasDisplayedError = false;
+    }
+  }
+
+  /**
    * Method to set validity values
    * @param newValue - valid new value
    */