+++ packages/stencil-helpers/src/components/index.ts
@@ -161,14 +161,21 @@ const getChildWindows = (localWindow: Window, windows: Window[] = []): Window[] 
 export const isValidString = (value: unknown): value is string => typeof value === 'string' && value.trim() !== '';
 
 /**
- * Value to stringify
- * @param value
+ * Stringify value
+ * @param value - value to stringify
  * @returns stringified value
  */
 export const toString = (value: unknown): string => {
   if (typeof value === 'object') return JSON.stringify(value);
   else return `${value}`;
 };
+
+/**
+ * Validate number
+ * @param value - value to check
+ * @returns `true` if number is valid
+ */
+export const isValidNumber = (value: unknown): value is number => typeof value === 'number' && !Number.isNaN(value);
 
 /**
  * Cleans string characters by removing special characters and converting to lowercase.