+++ tools/release-bot/bin/release.js
@@ -40,7 +40,6 @@ const converter = new showdown.Converter();
     }
   }
 
-  // TODO: Storybook is for now dedicated to mg-components but it is planned to make it the core-ui website.
   changelog.push(`\nView the complete changelog file here: [CHANGELOG](http://core.pages.mgdis.fr/core-documentation/docs/core-ui/CHANGELOG)`);
 
   const message = {