+++ configs/jest-config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # jest-config
 
+## 1.0.8
+
+### Patch Changes
+
+- f73a81d: Updated dependencies
+
 ## 1.0.7
 
 ### Patch Changes