+++ packages/linting-stencil/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @mgdis/linting-stencil
 
+## 1.3.2
+
+### Patch Changes
+
+- f73a81d: Updated dependencies
+
 ## 1.3.1
 
 ### Patch Changes